Add files via upload
Arrays in Java are used to store multiple values in a single variable, instead of declaring separate variables for each value. Here are some important points about arrays in Java:

Creating an Array
You can create an array by defining the type of its elements and the number of elements it will hold.